Update maintenance and cloudletstate handling to do FindCloudlet (#88)
* Update maintenance and cloudletstate handling to do FindCloudlet
* Bug: confusion between updates and num samples for Site test.
* Update docs with subsections.
* Add location permissions and section headers to documentation.
* Allow null sub configs for location and latency tests. Spec for maxNumberOfUpdates <= 0 now means infinity.
* Bump to 2.6.4

+++ EmptyMatchEngineApp/matchingengine/src/main/java/com/mobiledgex/matchingengine/edgeeventsconfig/ClientEventsConfig.java
@@ -20,7 +20,7 @@ package com.mobiledgex.matchingengine.edgeeventsconfig;
 public class ClientEventsConfig {
     public UpdatePattern updatePattern;
     public double updateIntervalSeconds; // in seconds
-    public int maxNumberOfUpdates;
+    public long maxNumberOfUpdates;
 
     public enum UpdatePattern {
         onStart,
@@ -31,7 +31,7 @@ public class ClientEventsConfig {
     public ClientEventsConfig() {
         updatePattern = UpdatePattern.onTrigger;
         updateIntervalSeconds = 30;
-        maxNumberOfUpdates = 0;
+        maxNumberOfUpdates = 0; // <= 0 means "infinity".
     }
 
     public ClientEventsConfig(ClientEventsConfig clientEventsConfig) {
